Enhance stdout logging
Previously logs were not being written to stdout because while we
were redirecting an existing log file to stdout, we were not accounting for vsftpd creating new log files when the user logs in. As a result I needed to include a new script that listens on the correct log files and tails them if requested.

This change also removes the hardcoded logging to stdout.